#105 파티 디테일 변경된 디자인 적용
