+++ src/components/Icon/svgs/index.ts
@@ -25,3 +25,4 @@ export { default as Pencil } from './pencil.svg?react'
 export { default as Remove } from './remove.svg?react'
 export { default as Camera } from './camera.svg?react'
 export { default as Check } from './check.svg?react'
+export { default as Copy } from './copy.svg?react'
