+++ src/pages/PartyDetailPage/components/PartyLocation.module.scss
@@ -1,0 +1,35 @@
+.Title {
+  font-size: rem(18px);
+  font-weight: 700;
+  color: $gray-900;
+  line-height: 1.7;
+}
+
+.KakaoMap {
+  width: 100%;
+  height: 150px;
+  cursor: pointer;
+  margin-top: 20px;
+}
+
+.Info {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  font-size: rem(16px);
+  margin-top: 16px;
+}
+
+.Name {
+  color: $gray-800;
+  font-weight: 600;
+}
+
+.SubInfo {
+  color: $gray-600;
+  font-size: rem(14px);
+  display: flex;
+  gap: 6px;
+  justify-content: start;
+  align-items: center;
+}
